pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = "Backend boilerplate for Django with DRF"
authors = ["JianxuanLi"]
license = "Private"

[tool.poetry.dependencies]
python = "3.8.10"
django-cors-headers = "^3.8.0"
coverage = "^5.5"
djangorestframework = "^3.12.4"
psycopg2-binary = "^2.9.1"
itsdangerous = "^2.0.1"
Pillow = "^8.3.1"
shortuuid = "^1.0.1"
GDAL = "3.0.4"
Shapely = "^1.7.1"
pandas = "^1.3.3"
DBUtils = "^2.0.2"
requests = "^2.26.0"
drf-yasg = "^1.20.0"
pyparsing = "3.0.0"
coreschema = "^0.0.4"
Django = "^3.2.9"
zeep = "^4.1.0"
django-filter = "2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
